Fixed one remaining 64-bit truncation issue missed in the last CL.
BUG=396
Review URL: https://codereview.appspot.com/7323055

git-svn-id: https://angleproject.googlecode.com/svn/branches/dx11proto@1961 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ src/compiler/glslang_tab.cpp
@@ -4479,7 +4479,7 @@ yyreduce:
         // knows where to find parameters.
         //
         TIntermAggregate* paramNodes = new TIntermAggregate;
-        for (int i = 0; i < function->getParamCount(); i++) {
+        for (size_t i = 0; i < function->getParamCount(); i++) {
             const TParameter& param = function->getParam(i);
             if (param.name != 0) {
                 TVariable *variable = new TVariable(param.name, *param.type);
@@ -4755,5 +4755,3 @@ yyreturn:
 int glslang_parse(TParseContext* context) {
     return yyparse(context);
 }
-
-
